refactor: make NotificationType type shared
